test: update Playwright tests for admin authentication and CRUD operations
- Modified Playwright test commands to run all tests in the e2e/admin directory.
- Added new tests for JWT expiration, login/logout functionality, and CRUD operations for posts and tweets.
- Implemented pagination tests for tweets, ensuring correct navigation and tweet count validation.
- Included checks for success messages and button states in the tests.
