+++ .github/workflows/playwright_admin.yml
@@ -36,7 +36,7 @@ jobs:
       - name: Build Frontend App
         run: pnpm build:e2e
       - name: Run Playwright tests
-        run: pnpm exec playwright test admin.spec.ts authentication.spec.ts
+        run: pnpm exec playwright test e2e/admin/
       - uses: actions/upload-artifact@v4
         if: always()
         with:
