+++ e2e/admin/post-crud.spec.ts
@@ -3,38 +3,12 @@ import util from 'node:util'
 
 import { expect } from '@playwright/test'
 
-import { test } from './helper'
+import { test } from '../helper'
 
 const exec = util.promisify(execCb)
 
 test.beforeAll(async () => {
   await exec('pnpm db:reset')
-})
-
-test.describe('login & logout', () => {
-  test('show login button', async ({ page }) => {
-    await page.goto('http://localhost:3000/')
-    await page.keyboard.press('x')
-    await expect(page.getByTestId('login-link')).toBeVisible()
-  })
-
-  test('failed login with incorrect user/password', async ({ page }) => {
-    await page.goto('http://localhost:3000')
-    await page.keyboard.press('x')
-    await page.click('[data-testid=login-link]')
-    await page.fill('[data-testid=name-input]', 'wefjweiofjwie')
-    await page.fill('[data-testid=password-input]', 'wfjweoifjio23r03')
-    await page.click('[data-testid=submit-btn]')
-    await expect(page.getByTestId('snackbar')).toHaveText('User does not exist')
-  })
-
-  test('successful Logout', async ({ authenticated: page }) => {
-    await page.goto('http://localhost:3000')
-    await page.keyboard.press('x')
-    await page.click('[data-testid=logout-link]')
-    await expect(page).toHaveURL('http://localhost:3000/')
-    await expect(page.getByTestId('login-link')).toBeVisible()
-  })
 })
 
 test.describe('CRUD post operation', () => {
@@ -106,62 +80,6 @@ test.describe('CRUD post operation', () => {
     await expect(page).toHaveURL('http://localhost:3000/dashboard')
     await expect(page.getByTestId('snackbar')).toHaveText('Delete Successful!')
     await expect(page.locator('main')).not.toHaveText(/Edit Title!/)
-  })
-})
-
-test.describe('Tweet CRUD', () => {
-  // TODO: add tweet CRUD test
-  test('create new tweet', async ({ authenticated: page }) => {
-    await page.goto('http://localhost:3000/')
-    await page.keyboard.press('x')
-    await page.click('[data-testid="tweet-link"]')
-    await expect(page).toHaveURL('http://localhost:3000/dashboard/tweet')
-
-    // Wait for the page to load
-    await expect(page.locator('h2')).toContainText('Tweets')
-
-    // Create a new tweet
-    const testTweetText = `Test tweet ${Date.now()}`
-    await page.fill('input[name="text"]', testTweetText)
-    await page.click('button[type="submit"]')
-
-    // Wait for success message using the correct snackbar test id
-    await expect(page.locator('[data-testid="snackbar"]')).toContainText(
-      'Tweet created successfully',
-    )
-
-    // Verify the tweet appears in the list automatically (without manual refresh)
-    await expect(page.locator(`text=${testTweetText}`)).toBeVisible()
-
-    // Verify the form was reset
-    await expect(page.locator('input[name="text"]')).toHaveValue('')
-  })
-
-  test('delete tweet', async ({ authenticated: page }) => {
-    await page.goto('http://localhost:3000/')
-    await page.keyboard.press('x')
-    await page.click('[data-testid="tweet-link"]')
-    await expect(page).toHaveURL('http://localhost:3000/dashboard/tweet')
-
-    // Wait for the page to load
-    await expect(page.locator('h2')).toContainText('Tweets')
-
-    // Wait for tweets to load
-    const tweetCards = page.locator('[data-testid^="tweet-card-"]')
-
-    // Get the first tweet and its delete button
-    const firstTweetCard = tweetCards.first()
-    const deleteButton = firstTweetCard.locator(
-      '[data-testid^="delete-tweet-"]',
-    )
-
-    // Click the delete button
-    await deleteButton.click()
-
-    // Wait for success message
-    await expect(page.locator('[data-testid="snackbar"]')).toContainText(
-      'Tweet deleted successfully',
-    )
   })
 })
